project/requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.11b1
click==8.0.3
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.0
